requirements.txt
asgiref==3.6.0
cloudinary==1.33.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-admin-rangefilter==0.10.0
django-allauth==0.54.0
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==1.14.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.6
pydantic==1.10.7
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
